backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
langchain==0.0.335
openai==1.3.5
pinecone-client==2.2.4
kafka-python==2.0.2
prometheus-client==0.19.0
python-dotenv==1.0.0
django-environ==0.11.2
gunicorn==21.2.0
uvicorn[standard]==0.24.0
websockets==12.0
daphne==4.0.0
django-extensions==3.2.3
django-filter==23.4
djangorestframework-simplejwt==5.3.0
Faker==20.1.0
pytest-django==4.7.0
pytest==7.4.3
factory-boy==3.3.0
